+++ src/nn/nn_ops.h
@@ -8,7 +8,7 @@ extern const struct nlop_s* nlop_rand_split_create(int N, const long dims[__VLA(
 extern const struct nlop_s* nlop_rand_mask_fixed_create(int N, const long dims[__VLA(N)], float p, unsigned long bat_flags);
 extern const struct nlop_s* nlop_rand_split_fixed_create(int N, const long dims[__VLA(N)], unsigned long shared_dims_flag, unsigned long bat_dims_flag, float p, unsigned long fix_flags, const _Complex float* _fix_first, float leaky_val);
 extern const struct nlop_s* nlop_maxpool_create(int N, const long dims[__VLA(N)], const long pool_size[__VLA(N)]);
-extern const struct nlop_s* nlop_dropout_create(int N, const long dims[__VLA(N)], float p, unsigned int shared_dims_flag);
+extern const struct nlop_s* nlop_dropout_create(int N, const long dims[__VLA(N)], float p, unsigned long shared_dims_flag);
 extern const struct nlop_s* nlop_noise_create(int N, const long dims[__VLA(N)], float sigma, unsigned long shared_dims_flag, unsigned long shared_sigma_flag);
 extern const struct nlop_s* nlop_add_noise_create(int N, const long dims[__VLA(N)], float sigma, unsigned long shared_dims_flag, unsigned long shared_sigma_flag);
